refactor(loading button): shift to ts
