+++ src/components/LoadingButton/index.ts
@@ -1,0 +1,1 @@
+export * from "./LoadingButton"
